% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStatistical.R
\name{bootstrapPrecisionMeasures}
\alias{bootstrapPrecisionMeasures}
\alias{bootstrapSE.mean}
\alias{bootstrapCI.mean}
\alias{bootstrapSE.median}
\alias{bootstrapCI.median}
\alias{bootstrapSE.hmean}
\alias{bootstrapCI.hmean}
\alias{bootstrapSE.gmean}
\alias{bootstrapCI.gmean}
\alias{bootstrapSE.var}
\alias{bootstrapCI.var}
\alias{bootstrapSE.sd}
\alias{bootstrapCI.sd}
\title{Bootstrapped measures of precision}
\usage{
bootstrapSE.mean(x)

bootstrapCI.mean(x, gamma)

bootstrapSE.median(x)

bootstrapCI.median(x, gamma)

bootstrapSE.hmean(x)

bootstrapCI.hmean(x, gamma)

bootstrapSE.gmean(x)

bootstrapCI.gmean(x, gamma)

bootstrapSE.var(x)

bootstrapCI.var(x, gamma)

bootstrapSE.sd(x)

bootstrapCI.sd(x, gamma)
}
\arguments{
\item{x}{a vector of numbers, the sample data (mandatory);}

\item{gamma}{a confidence level for CI (default 0.95).}
}
\value{
a measure of precision (SE) or an interval of precision (CI).
}
\description{
superb also comes with a few built-in measures of
precisions that uses bootstrap. More can be added based on users needs.
All \code{bootstrapSE.fct()} functions produces an interval width;
all \code{bootstrapCI.fct()} produces the lower and upper limits of an interval.
These estimates are based on 5,000 sub-samples by default. Change this
default with\code{options("superb.bootstrapIter" = number )}.
See \insertCite{et94}{superb} for a comprehensive introduction.
All "superbPlot-compatible" precision measures must have these parameters:
}
\examples{
# the confidence interval of the mean for default 95\% and 90\% confidence level
bootstrapCI.mean( c(1,2,3) )
bootstrapCI.mean( c(1,2,3), gamma = 0.90)

# Standard errors for standard deviation or variance
bootstrapSE.sd( c(1,2,3) )
bootstrapSE.var( c(1,2,3) )

}
\references{
\insertAllCited{}
}
