% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFigure3.R
\docType{data}
\name{dataFigure3}
\alias{dataFigure3}
\title{Data for Figure 3}
\format{
An object of class data.frame.
}
\source{
\href{https://doi.org/10.1037/met0000055}{doi}
}
\usage{
data(dataFigure3)
}
\description{
The data, inspired from Cousineau and Laurencelle, 2016, is an example where the
"stand-alone" 95% confidence interval of the means returns
a result in contradiction with the result of a statistical test.
The paradoxical result is resovled by using adjusted confidence intervals,
here the cluster- and different-adjusted confidence interval.
}
\examples{
data(dataFigure3)

## realize the plot with unadjusted (left) and ajusted (right) 95\% confidence intervals
plt3a <- superbPlot(dataFigure3, BSFactor = "grp", 
    adjustments=list(purpose = "difference", samplingDesign = "SRS"),
    variables = c("VD"), plotStyle="bar" ) + 
  xlab("Group") + ylab("Score") + labs(title="Difference-adjusted 95\% CI\n") +
  coord_cartesian( ylim = c(85,115) ) +
  geom_hline(yintercept = 100, colour = "black", size = 0.5, linetype=2)
plt3b <- superbPlot(dataFigure3, BSFactor = "grp", 
    adjustments=list(purpose = "difference", samplingDesign = "CRS"),
    variables = c("VD"), plotStyle="bar", clusterColumn = "cluster" ) + 
  xlab("Group") + ylab("Score") + labs(title="Cluster and difference-adjusted\n95\% CI") +
  coord_cartesian( ylim = c(85,115) ) + 
  geom_hline(yintercept = 100, colour = "black", size = 0.5, linetype=2)
plt3  <- grid.arrange(plt3a,plt3b,ncol=2)

## realise the correct t-test to see the discrepancy
res   <- t.test(dataFigure3$VD[dataFigure3$grp==1], 
               dataFigure3$VD[dataFigure3$grp==2],
               var.equal=T)
micc  <- mean(c(0.491334683772226, 0.20385744842838)) # mean ICC given by superbPlot
lam   <- lambda(c(micc, 5,5,5,5,5,5))
tcorr <- res$statistic / lam
pcorr <- 1-pt(tcorr,4)#' 

}
\references{
Cousineau, D., & Laurencelle, L. (2016) Psychological Methods, 21:121-135
(\href{https://doi.org/10.1037/met0000055}{doi})
}
\keyword{datasets}
