% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon.R
\name{epsilon}
\alias{epsilon}
\title{epsilon}
\usage{
epsilon(dta, cols)
}
\arguments{
\item{dta}{a data.frame}

\item{cols}{a vector of column names indicating the relevant
columns on which to compute epsilon. Any other columns are ignored.}
}
\value{
returns the Hyunh-Feldt estimate of sphericity epsilon
}
\description{
epsilon is a measure of sphericity created by 
Greenhouse and Geiser (1958). The original measure was biased
and therefore, Hyunh and Feldt (1976) produced a revised version
(note that the 1976 paper contained typos that were uncorrected
in SPSS; Lecoutre, 1991)
}
