% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyunhFeldtEpsilon.R
\name{HyunhFeldtEpsilon}
\alias{HyunhFeldtEpsilon}
\title{Hyunh and Feldt's epsilon measure of sphericity}
\usage{
HyunhFeldtEpsilon(dta, cols)
}
\arguments{
\item{dta}{a data.frame}

\item{cols}{a vector of column names indicating the relevant
columns on which to compute epsilon. Any other columns are ignored.}
}
\value{
returns the Hyunh-Feldt estimate of sphericity epsilon
}
\description{
HyunhFeldtEpsilon() is a measure of sphericity created by
\insertCite{gg58;textual}{superb}. The original measure was biased
and therefore, \insertCite{hf76;textual}{superb} produced a revised version
(note that the 1976 paper contained typos that were uncorrected
in SPSS; \insertCite{l91;textual}{superb})
}
\references{
\insertAllCited{}
}
