% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotESP.R
\name{plotESP}
\alias{plotESP}
\title{plotESP}
\usage{
plotESP(data, bsFactor = NULL, wsFactor = NULL, factorOrder, variables,
  statistic = "mean", errorbar = "CI", gamma = 0.95,
  adjustments = list(purpose = "single", popSize = Inf, decorrelation =
  "none", samplingScheme = "SRS"), showPlot = TRUE, plotStyle = "bar",
  errorParams = list(width = 0.8), pointParams = list(),
  Debug = FALSE, preprocessfct = NULL, postprocessfct = NULL)
}
\arguments{
\item{data}{Dataframe in wide format}

\item{bsFactor}{The name of the columns containing the between-subject factor(s)}

\item{wsFactor}{The name of the within-subject factor(s)}

\item{factorOrder}{Order of factors as shown in the graph (x axis, groups, horizontal 
panels, vertical panels)}

\item{variables}{The dependent variable(s)}

\item{statistic}{The summary statistic function to use}

\item{errorbar}{The function that computes the error bar. Should be "CI" or "SE" or 
any function name. Defaults to "SE"}

\item{gamma}{The converage factor; necessary when errorbar == "CI". Default is 0.95.}

\item{adjustments}{List of adjustments as described below:}

\item{showPlot}{Defaults to TRUE. Set to FALSE if you want the output to be the summary statistics and intervals.}

\item{plotStyle}{The type of object to plot on the graph. Can be either "bar" or "line".
Defaults to "bar".}

\item{errorParams}{a list of ggplot2 parameters for geom_errobar (see ?geom_errorbar)}

\item{pointParams}{a list of ggplot2 parameters to input inside geoms (see ?geom_bar2)}

\item{Debug}{export internal information into global environment. Default is FALSE}

\item{preprocessfct}{is a transform (or vector of) to be performed first on data matrix of each group}

\item{postprocessfct}{is a transform (or vector of)}

\item{adjustments$popsize}{Size of the population under study. Defaults to Inf}

\item{adjustments$purpose}{The purpose of the comparisons. Defaults to "single". 
Can be "single" or "difference".}

\item{adjustments$decorrelation}{Decorrelation method for repeated measure designs. 
     Chooses among the methods ("CM", "LM", "CA" or "none"). Defaults to "none".
Default is adjustments = list(purpose = "single", popSize = Inf, decorrelation = "none")}
}
\value{
a plot with the correct error bars or a table of those summary statistics.
        The plot is a ggplot2 object with can be modified with additional declarations.
}
\description{
Plots standard error or confidence interval for various descriptive 
     statistics under various designs, sampling schemes, population size and purposes,
     according to the ESP framework.
}
\examples{
# basic example using a built-in dataframe as data; 
# by default, the mean is computed and the error bar are 95\% confidence intervals
plotESP(ToothGrowth, bsFactor = c("dose", "supp"), 
  variables = "len") 

# exampled changing the summary statistics to the median and
# the error bar to 90\% confidence intervals
plotESP(ToothGrowth, bsFactor = c("dose", "supp"), 
  variables = "len", statistic = "median", errorbar = "CI", gamma = .90) 

# exampled introducing adjustments
# for pairwise comparisons and assuming that the whole population is limited to 200 persons
plotESP(ToothGrowth, bsFactor = c("dose", "supp"), 
  variables = "len",  
  adjustments = list( purpose = "difference", popSize = 200) )

# This example add ggplots directives to the plot produced
plotESP(ToothGrowth, bsFactor = c("dose", "supp"), 
  variables = "len") + 
  xlab("Dose") + ylab("Tooth Growth") +
  theme_bw()

}
