% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTransformation.R
\name{poolSDTransform}
\alias{poolSDTransform}
\title{poolSDTransform}
\usage{
poolSDTransform(dta, variables)
}
\arguments{
\item{dta}{a data.frame containing the data in wide format;}

\item{variables}{a vector of column names on which the transformation will be applied.
the remaining columns will be left unchanged}
}
\value{
a data.frame of the same form as dta with the variables transformed.
}
\description{
poolSDTransform  is a 
transformations that can be applied to a matrix of data.
}
