% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDebug.R
\name{runDebug}
\alias{runDebug}
\title{runDebug}
\usage{
runDebug(state, title, vars, vals)
}
\arguments{
\item{state}{sd fasdf}

\item{title}{sadfasdf}

\item{vars}{asdfasdf}

\item{vals}{asdfasdf}
}
\value{

}
\description{
Plots standard error or confidence interval for various descriptive 
     statistics under various designs, sampling schemes, population size and purposes,
     according to the ESP framework.
}
