% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTransformation.R
\name{subjectCenteringTransform}
\alias{subjectCenteringTransform}
\title{subjectCenteringTransform}
\usage{
subjectCenteringTransform(dta, variables)
}
\arguments{
\item{dta}{a data.frame containing the data in wide format;}

\item{variables}{a vector of column names on which the transformation will be applied.
the remaining columns will be left unchanged}
}
\value{
a data.frame of the same form as dta with the variables transformed.
}
\description{
subjectCenteringTransform is a 
transformations that can be applied to a matrix of data.
}
