% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superbShiny.R
\name{superbShiny}
\alias{superbShiny}
\title{User Interface to get summary plot of any statistics with adjusted error bars.}
\usage{
superbShiny()
}
\value{
A plot that can be cut-and-paste.
}
\description{
The function \code{suberbShiny()} provides a simple user interface
to plot standard error or confidence interval for various descriptive
statistics under various designs, population size and purposes,
according to the \code{suberb} framework. See \insertCite{cgh21}{superb} for more.
Also see this \insertCite{w21}{youtube} presentation for a demo using
the shinyapps.io installation from https://dcousin3.shinyapps.io/superbshiny/
Limitations: it is not possible to use custom-made statistics with the
graphical user interface, nor is it possible to request an adjustment for cluster-
randomized sampling. These options are available with \code{superbPlot()}.
}
\examples{
# Launch the user interface: 
\donttest{
if (interactive())
   superbShiny() 
}

}
\references{
\insertAllCited{}
}
