% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superbPlot.R
\name{two_step_transform}
\alias{two_step_transform}
\alias{subject_centering_transform}
\alias{bias_correction_transform}
\alias{pool_sd_transform}
\title{transformations}
\usage{
two_step_transform(dta, variables)
}
\arguments{
\item{dta}{a data.frame containing the data in wide format;}

\item{variables}{a vector of column names on which the transformation will be applied.
the remaining columns will be left unchanged}
}
\value{
a data.frame of the same form as dta with the variables transformed.
}
\description{
two_step_transform, subject_centering_transform, 
bias_correction_transform and pool_sd_transform are four 
transformations that can be applied to a matrix of data.
}
