% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTransformation.R
\name{twoStepTransform}
\alias{twoStepTransform}
\title{two-step transform for subject centering and bias correction}
\usage{
twoStepTransform(dta, variables)
}
\arguments{
\item{dta}{a data.frame containing the data in wide format;}

\item{variables}{a vector of column names on which the transformation will be applied.
the remaining columns will be left unchanged}
}
\value{
a data.frame of the same form as dta with the variables transformed.

This function is useful when passed to the argument \code{preprocessfct} of \code{superb()}
where it performs a modification of the data matrix.
}
\description{
\code{twoStepTransform} is a transformation that can
be applied to a matrix of data. The resulting matrix is both
subject-centered and bias corrected, a technique called
the CM technique \insertCite{b12,c05,m08}{superb}
}
\references{
\insertAllCited{}
}
