% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTransformation.R
\name{poolSDTransform}
\alias{poolSDTransform}
\title{pooled standard deviation transform}
\usage{
poolSDTransform(dta, variables)
}
\arguments{
\item{dta}{a data.frame containing the data in wide format;}

\item{variables}{a vector of column names on which the transformation will be applied.
the remaining columns will be left unchanged}
}
\value{
a data.frame of the same form as dta with the variables transformed.

This function is useful when passed to the argument \code{preprocessfct} of \code{superbPlot()}
where it performs a modification of the data matrix.
}
\description{
\code{poolSDTransform}  is a transformations that can
be applied to a matrix of data. The resulting matrix has the column-
standard deviations equal to the pool standard deviations of the
individual columns, the solution adopted by \insertCite{lm94}{superb}.
}
\references{
\insertAllCited{}
}
