% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDebug.R
\name{runDebug}
\alias{runDebug}
\title{runDebug}
\usage{
runDebug(state, title, vars, vals)
}
\arguments{
\item{state}{boolean (TRUE to activate runDebug)}

\item{title}{string text to be displayed when this function is triggered}

\item{vars}{strings names of the variables to be placed in the globalenvironment}

\item{vals}{numeric values to be given to the variables.}
}
\value{

}
\description{
runDebug is an internal function used by GRD and superbPlot
     to help in debugging the functions. It assigns in the global environment
     the variables that are local in a function so that they become visible.
}
