% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTransformation.R
\name{subjectCenteringTransform}
\alias{subjectCenteringTransform}
\title{subject-centering transform}
\usage{
subjectCenteringTransform(dta, variables)
}
\arguments{
\item{dta}{a data.frame containing the data in wide format;}

\item{variables}{a vector of column names on which the transformation will be applied.
the remaining columns will be left unchanged}
}
\value{
a data.frame of the same form as dta with the variables transformed.

This function is useful when passed to the argument \code{preprocessfct} of \code{superb()}
where it performs a modification of the data matrix.
}
\description{
\code{subjectCenteringTransform} is a transformation that can
be applied to a matrix of data. the resulting matrix have means
that are centered on the grand mean, subject-wise \insertCite{c05}{superb}.
}
\references{
\insertAllCited{}
}
