% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suberb.R
\docType{package}
\name{superb-package}
\alias{superb}
\alias{superb-package}
\title{superb: Computes standard error and confidence interval of means under various designs and sampling schemes}
\description{
Computes standard error and confidence interval of means under various designs and sampling schemes
}
\details{
suberb is a library to perform descriptive statistics plots
based on the ESP framework.
The unique function is

plotESP(df,  ...)  

where df is a dataframe.

See help(plotESP) for more details.
}
\author{
\strong{Maintainer}: Denis Cousineau \email{denis.cousineau@uottawa.ca}

Authors:
\itemize{
  \item Felix Chiasson \email{felix@chiasson.co}
}

}
\keyword{internal}
